requirements.txt
connexion==2.6.0
connexion[swagger-ui] == 2.6.0
setuptools >= 21.0.0
decorator==4.4.2
dill==0.3.2
future==0.18.2
joblib==0.16.0
networkx==2.4
numpy==1.19.1
pandas==1.1.1
protobuf==3.13.0
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
snorkel-metal==0.5.0
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.6.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.6.0
decorator==4.4.2
dill==0.3.2
Flask==1.1.2
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
MarkupSafe==1.1.1
networkx==2.4
numpy==1.19.1
openapi-spec-validator==0.2.9
pandas==1.1.1
protobuf==3.13.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scikit-learn==0.23.2
scipy==1.5.2
sentencepiece==0.1.91
six==1.15.0
snorkel-metal==0.5.0
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.6.0
torchtext==0.7.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
absl-py==0.10.0
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.6.0
decorator==4.4.2
dill==0.3.2
Flask==1.1.2
future==0.18.2
grpcio==1.31.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
jsonschema==3.2.0
Markdown==3.2.2
MarkupSafe==1.1.1
munkres==1.1.2
numpy==1.19.1
openapi-spec-validator==0.2.9
pandas==1.1.1
protobuf==3.13.0
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
scipy==1.5.2
sentencepiece==0.1.91
six==1.15.0
snorkel==0.9.6
snorkel-metal==0.5.0
tensorboard==1.15.0
tensorboardX==2.1
threadpoolctl==2.1.0
torch==1.6.0
torchtext==0.7.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.1.0
